refactor: isElementClamped를 shared/lib로 이동 + 도달 불가 canExpand 가드 제거 (PR #627 리뷰)
