+++ src/shared/lib/clamp.ts
