+++ src/widgets/home/hooks/useIsClamped.ts
@@ -1,7 +1,7 @@
 'use client';
 
 import { useEffect, useRef, useState, type RefObject } from 'react';
-import { isElementClamped } from '../utils/clamp';
+import { isElementClamped } from '@/shared/lib/clamp';
 
 interface UseIsClampedReturn {
     ref: RefObject<HTMLParagraphElement | null>;
